+++ src/models/CNNRGB/CNNRGB-standing_laying/q4-rgb-standing-laying/src/preprocessing.py
@@ -1,0 +1,33 @@
+import os
+
+import tensorflow as tf
+
+from config import CONFIG
+
+class_names = ['laying', 'standing']
+
+
+def get_label(file_path):
+    # convert the path to a list of path components
+    parts = tf.strings.split(file_path, os.path.sep)
+    # The second to last is the class-directory
+    one_hot = parts[-2] == class_names
+    one_hot = tf.cast(one_hot, tf.int64)
+    # Integer encode the label
+    return tf.argmax(one_hot)
+
+
+def decode_img(img):
+    # convert the compressed string to a 3D uint8 tensor
+    img = tf.image.decode_jpeg(img, channels=3)
+    img = tf.cast(img, tf.float32) * (1. / CONFIG.NORMALIZATION_VALUE)
+    # resize the image to the desired size
+    return tf.image.resize(img, [CONFIG.IMAGE_TARGET_HEIGHT, CONFIG.IMAGE_TARGET_WIDTH])
+
+
+def process_path(file_path):
+    label = get_label(file_path)
+    # load the raw data from the file as a string
+    img = tf.io.read_file(file_path)
+    img = decode_img(img)
+    return img, label
