+++ src/models/CNNRGB/CNNRGB-standing_laying/q4-rgb-standing-laying/src/test/test_train.py
@@ -5,13 +5,12 @@ import numpy as np
 import tensorflow as tf
 
 sys.path.append(str(Path(__file__).parents[1]))
-from train import process_path
 from constants import REPO_DIR
-print(REPO_DIR)
+from preprocessing import process_path
 
 
 def test_get_label_0():
-    paths = [str(REPO_DIR)+'/data/anon-rgb-classification/test/laying/rgb_1597886481-znzw5yzhjh_1597886481929_200_13520.886623759001.jpg']
+    paths = [str(REPO_DIR) + '/src/models/CNNRGB/CNNRGB-standing_laying/q4-rgb-standing-laying/src/test/anon-rgb-classification/test/laying/rgb_1597886481-znzw5yzhjh_1597886481929_200_13520.886623759001.jpg']
     dataset = tf.data.Dataset.from_tensor_slices(paths)
     dataset = dataset.map(lambda path: process_path(path))
 
@@ -20,7 +19,7 @@ def test_get_label_0():
 
 
 def test_get_label_1():
-    paths = [str(REPO_DIR)+'/data/anon-rgb-classification/test/standing/rgb_1585352016-s51bhrzmtt_1592712865086_100_520010.772824759.jpg']
+    paths = [str(REPO_DIR) + '/src/models/CNNRGB/CNNRGB-standing_laying/q4-rgb-standing-laying/src/test/anon-rgb-classification/test/standing/rgb_1585352016-s51bhrzmtt_1592712865086_100_520010.772824759.jpg']
     dataset = tf.data.Dataset.from_tensor_slices(paths)
     dataset = dataset.map(lambda path: process_path(path))
 
@@ -29,10 +28,10 @@ def test_get_label_1():
 
 
 def test_data():
-    paths = [str(REPO_DIR)+'/data/anon-rgb-classification/test/laying/rgb_1597886481-znzw5yzhjh_1597886481929_200_13520.886623759001.jpg']
+    paths = [str(REPO_DIR) + '/src/models/CNNRGB/CNNRGB-standing_laying/q4-rgb-standing-laying/src/test/anon-rgb-classification/test/laying/rgb_1597886481-znzw5yzhjh_1597886481929_200_13520.886623759001.jpg']
     dataset = tf.data.Dataset.from_tensor_slices(paths)
     dataset = dataset.map(lambda path: process_path(path))
 
     for a, _ in dataset.as_numpy_iterator():
         assert np.max(a) <= 1
-        assert np.min(a) >= 0 and np.min(a) >= 0
+        assert np.min(a) >= 0
